backend/requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask-cors
python-dotenv
pymysql
cryptography
